requirements.txt
Flask==3.1.2 \
    --hash=sha256:3aa6738a0905a8533daa08bd8a2a74041a275e3ad6f1e21cd998ca55a17ec5b7 \
    --hash=sha256:d71e67ae89310f0dbab322d56c07e2f7f4a0d9c5a4ea6eedb1f3cac3aa2288e7

Pillow==11.3.0 \
    --hash=sha256:5fc67b31cc66b8aad5fb427ecf0b37c823c9b94a8f6d3c9ac0e23affd5afdbcb \
    --hash=sha256:ab21c06e498fb021c1df6dde9146a33ed5bf77029d5cbdb7172587beafb53f72

gunicorn==23.0.0 \
    --hash=sha256:92c001a71f08c3f44bca5ef54e9fb9cc1669ac6e8d5c80b4bad8d94040364e0d \
    --hash=sha256:ac451cc735987c30e86282f76c84b8373b0a090ade27bf8c6454899e6472ab91